+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add - Demo screenshot.
 * Add - Sponsor link to original Intelephense project.
+* Add - Support workspace specific stubs.
 * Tweak - Update Intelephense to version 1.8.0.
 
 ## Version 1.0.0 - November, 9 2021